+++ index.js
@@ -1,14 +1,3 @@
-var github = require("./lib/github");
-fs = require("fs");
+var app = require("./app.js");
 
-var commits = github.githubCommits(
-  "GeneTinderholm",
-  "project_github_commit_feed"
-);
-commits
-  .then(data => {
-    console.log(data);
-  })
-  .catch(err => {
-    console.error(err);
-  });
+app.init();
